@@ -14,15 +14,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "17.3.12",
-    "@angular/core": "17.3.12",
-    "@angular/forms": "^17.3.12",
+    "@angular/animations": "^18.2.13",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "18.2.13",
+    "@angular/core": "18.2.13",
+    "@angular/forms": "^18.2.13",
     "@angular/http": "7.2.16",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
     "core-js": "^2.4.1",
     "ng-click-outside": "^4.0.0",
     "ngx-device-detector": "^4.0.1",
@@ -34,7 +34,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^15.0.4",
     "@angular/cli": "^15.0.4",
-    "@angular/compiler-cli": "^17.3.12",
+    "@angular/compiler-cli": "^18.2.13",
     "@types/jasmine": "2.5.53",
     "@types/node": "~8.0.16",
     "breakpoint-sass": "^2.7.1",
